+++ src/main/java/org/fasttrackit/DogFood.java
@@ -3,7 +3,7 @@ package org.fasttrackit;
 public class DogFood {
     Double price;
     Double quantity;
-    Double ageSelection;
+    int ageSelection;
     String flavor;
     String stature;
     String type;
