+++ src/main/java/org/fasttrackit/Vet.java
@@ -4,7 +4,7 @@ public class Vet {
     String name;
     String specialization;
     Double price;
-    Double experienceYears;
+    int experienceYears;
 
 
 }
